Made transaction work
