+++ package.json
@@ -32,11 +32,14 @@
   },
   "dependencies": {
     "@noble/hashes": "^1.1.5",
+    "@noble/secp256k1": "^1.7.0",
     "@popperjs/core": "^2.11.6",
-    "binstring": "^0.2.1",
+    "biginteger": "^1.0.3",
     "bip39": "^3.0.4",
     "bootstrap": "^5.2.3",
     "bootstrap-sass": "^3.4.3",
+    "bs58": "^5.0.0",
+    "buffer": "^6.0.3",
     "clean-webpack-plugin": "^4.0.0",
     "hdkey": "^2.0.1",
     "html-loader": "^4.2.0",
