+++ scripts/global.js
@@ -1,19 +1,21 @@
 import { Mempool } from "./mempool.js";
+import Masternode from "./masternode.js";
 import { en_translation } from "../locale/en/translation.js";
 import { uwu_translation } from "../locale/uwu/translation.js";
-import { translate } from "./i18n.js";
+import { translate, ALERTS } from "./i18n.js";
 import * as jdenticon from "jdenticon";
 import { masterKey, hasEncryptedWallet, importWallet, getNewAddress, isYourAddress } from "./wallet.js";
-import { submitAnalytics, networkEnabled, getBlockCount, arrRewards } from "./network.js";
+import { submitAnalytics, networkEnabled, getBlockCount, arrRewards, getStakingRewards } from "./network.js";
 import { start as settingsStart, cExplorer } from "./settings.js";
-import { createAlert, confirmPopup } from "./misc.js";
+import { createAlert, confirmPopup, sanitizeHTML } from "./misc.js";
+import { cChainParams, COIN } from "./chain_params.js";
 
 // TRANSLATION
 //Create an object of objects filled with all the translations
-const translatableLanguages = {
+export const translatableLanguages = {
     "en": en_translation,
     "uwu": uwu_translation
-}
+};
 
 export let doms = {};
 export let translation = {};
@@ -767,140 +769,6 @@ function createAlertWithFalse() {
     return false;
 }
 
-function validateAmount(nAmountSats, nMinSats = 10000) {
-    // Validate the amount is a valid number, and meets the minimum (if any)
-    if (nAmountSats < nMinSats || isNaN(nAmountSats))
-        return createAlertWithFalse('warning', ALERTS.INVALID_AMOUNT + ALERTS.VALIDATE_AMOUNT_LOW, [{"minimumAmount" : (nMinSats / COIN)}, {"coinTicker" : cChainParams.current.TICKER}], 2500);
-
-    // Validate the amount in Satoshi terms meets the coin's native decimal depth
-    if (!Number.isSafeInteger(nAmountSats))
-        return createAlertWithFalse('warning', ALERTS.INVALID_AMOUNT + '<br>' + ALERTS.VALIDATE_AMOUNT_DECIMAL,[{"coinDecimal" : COIN_DECIMALS}], 2500);
-
-    // Amount looks valid!
-    return true;
-}
-
-function undelegateGUI() {
-    if (masterKey.isViewOnly) {
-	return createAlert("warning", "Attempting to undelegate in view only mode.", 6000);
-    }
-    // Verify the amount
-    const nAmount = Math.round(Number(domGuiUndelegateAmount.value.trim()) * COIN);
-    if(!validateAmount(nAmount)) return;
-
-    undelegate(nAmount);
-}
-
-async function undelegate(nValue) {
-    if (!hasWalletUnlocked(true)) return;
-    
-    // Construct a TX and fetch Cold inputs
-    const nBalance = getStakingBalance();
-    const cTx = bitjs.transaction();
-    const cCoinControl = chooseUTXOs(cTx, nValue, 0, true);
-    if (!cCoinControl.success) return alert(cCoinControl.msg);
-
-    // Compute fee and change (or lack thereof)
-    const nFee = getFee(cTx.serialize().length);
-    const nChange = cCoinControl.nValue - (nFee + nValue);
-    const fReDelegateChange = nChange > 1.01 * COIN;
-    let reDelegateAddress;
-    let reDelegateAddressPath;
-    if (fReDelegateChange) {
-        // Enough change to resume cold staking, so we'll re-delegate change to the cold staking address
-        // Ensure the user has an address set - if not, request one!, Sanity
-        if (!cachedColdStakeAddr || cachedColdStakeAddr.length !== 34 || !cachedColdStakeAddr.startsWith(cChainParams.current.STAKING_PREFIX)) {
-            askForCSAddr(true);
-            return createAlert('success', ALERTS.SUCCESS_STAKING_ADDR ,[]);
-        }
-        // The re-delegated change output
-        [reDelegateAddress,reDelegateAddressPath] = await getNewAddress();
-        cTx.addcoldstakingoutput(reDelegateAddress, cachedColdStakeAddr, nChange / COIN);
-        console.log('Re-delegated delegation spend change!');
-    } else {
-        // Not enough change to cold stake, so we'll just unstake everything (and deduct the fee from the value)
-        nValue -= nFee;
-        console.log('Spent all CS dust into redeem address!');
-    }
-
-    const [outputKey,outputKeyPath] = await getNewAddress();
-    // The primary Cold-to-Public output
-    cTx.addoutput(outputKey, nValue / COIN);
-
-    // Debug-only verbose response
-    if (debug)doms.domHumanReadable.innerHTML = "Balance: " + (nBalance / COIN) + "<br>Fee: " + (nFee / COIN) + "<br>To: " + outputKey + "<br>Sent: " + (nValue / COIN) + (nChange > 0 ? "<br>Change Address: " + (fReDelegateChange ? cachedColdStakeAddr : outputKey) + "<br>Change: " + (nChange / COIN) : "");
-
-    if (hasHardwareWallet()) {
-        // Format the inputs how the Ledger SDK prefers
-        const arrInputs = [];
-        const arrAssociatedKeysets = [];
-        for (const cInput of cTx.inputs) {
-            const cInputFull = await getTxInfo(cInput.outpoint.hash);
-            arrInputs.push([await cHardwareWallet.splitTransaction(cInputFull.hex), cInput.outpoint.index]);
-            arrAssociatedKeysets.push(cInput.path);
-        }
-
-        // Construct the Ledger transaction
-        const cLedgerTx = await cHardwareWallet.splitTransaction(cTx.serialize());
-        const strOutputScriptHex = await cHardwareWallet
-              .serializeTransactionOutputs(cLedgerTx)
-              .toString("hex");
-
-        // Sign the transaction via Ledger
-        createAlert("info", ALERTS.CONFIRM_UNSTAKE_H_WALLET, [{"strHardwareName" : strHardwareName}], 7500);
-        const cLedgerSignedTx = await cHardwareWallet.createPaymentTransactionNew({
-            inputs: arrInputs,
-            associatedKeysets: arrAssociatedKeysets,
-            outputScriptHex: strOutputScriptHex,
-        });
-        const nInputLen = cTx.inputs.length;
-
-        // Put public key bytes instead of [3,195,174...]
-        const arrSignedTxBytes = Crypto.util.hexToBytes(cLedgerSignedTx);
-        const arrPubkey = findCompressedPubKey(arrSignedTxBytes);
-        const arrPubkeyWithScriptLen = addScriptLength(arrSignedTxBytes, arrPubkey, nInputLen);
-        const arrPubkeyWithScript = addExtraBytes(arrPubkeyWithScriptLen, arrPubkey, nInputLen);
-
-        const strSerialisedTx = Crypto.util.bytesToHex(arrPubkeyWithScript);
-
-        // Broadcast the Hardware (Ledger) TX
-        const result = await sendTransaction(strSerialisedTx, "<b>Delegation successfully spent!</b>");
-        if(result){
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            // Add our undelegation + change re-delegation (if any) to the local mempool
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(strSerialisedTx)))))));
-            if(fReDelegateChange){
-		mempool.addUTXO({id: futureTxid,path: reDelegateAddressPath,script:Crypto.util.bytesToHex(cTx.outputs[0].script) ,sats: nChange,vout: 0,status: Mempool.PENDING_COLD});
-		mempool.addUTXO({id: futureTxid,path: outputKeyPath,script:Crypto.util.bytesToHex(cTx.outputs[1].script) ,sats: nValue,vout: 1,status: Mempool.PENDING});
-            }else{
-		mempool.addUTXO({id: futureTxid,path: outputKeyPath,sats: nValue,script:Crypto.util.bytesToHex(cTx.outputs[0].script),vout: 0,status: Mempool.PENDING});
-            }
-        }
-    } else {
-        let sign= await cTx.sign(masterKey, 1, 'coldstake');
-        // Broadcast the software TX
-	
-        const result = await sendTransaction(sign, "<b>Delegation successfully spent!</b>");
-        if(result){
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            // Add our undelegation + change re-delegation (if any) to the local mempool
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(sign)))))));
-            if(fReDelegateChange){
-		mempool.addUTXO({id: futureTxid,path:reDelegateAddressPath,script:Crypto.util.bytesToHex(cTx.outputs[0].script) ,sats: nChange,vout: 0,status: Mempool.PENDING_COLD});
-		mempool.addUTXO({id: futureTxid,path: outputKeyPath,script:Crypto.util.bytesToHex(cTx.outputs[1].script) ,sats: nValue,vout: 1,status: Mempool.PENDING});
-            }else{
-		mempool.addUTXO({id: futureTxid,path: outputKeyPath,sats: nValue,script:Crypto.util.bytesToHex(cTx.outputs[0].script),vout: 0,status: Mempool.PENDING});
-            }
-        }
-    }
-
-  doms.domGenIt.innerHTML = "Continue";
-}
-
 function askForCSAddr(force = false) {
     if (force) cachedColdStakeAddr = null;
     if (cachedColdStakeAddr === "" || cachedColdStakeAddr === null) {
@@ -912,181 +780,6 @@ function askForCSAddr(force = false) {
     return false;
 }
 
-function delegateGUI() {
-    if(masterKey.isViewOnly) {
-	return createAlert("warning", "Attempting to delegate in view only mode.", 6000);
-    }
-    // Verify the amount; Delegations must be a minimum of 1 PIV, enforced by the network
-    const nAmount = Number(domGuiDelegateAmount.value.trim()) * COIN;
-    if (!validateAmount(nAmount, COIN)) return;
-
-    // Ensure the user has an address set - if not, request one!
-    if (!askForCSAddr()) return;      
-
-    // Sanity
-    if (cachedColdStakeAddr.length !== 34 || !cachedColdStakeAddr.startsWith(cChainParams.current.STAKING_PREFIX)) {
-        askForCSAddr(true);
-        return createAlert('success', ALERTS.SUCCESS_STAKING_ADDR_SET, []);
-    }
-    delegate(nAmount, cachedColdStakeAddr);
-}
-
-async function delegate(nValue, coldAddr) {
-    if (!hasWalletUnlocked(true)) return;
-
-    // Construct a TX and fetch Standard inputs
-    const nBalance = getBalance();
-    const cTx = bitjs.transaction();
-    const cCoinControl = chooseUTXOs(cTx, nValue, 0, false);
-    if (!cCoinControl.success) return alert(cCoinControl.msg);
-
-    // Compute fee and change (or lack thereof)
-    const nFee = getFee(cTx.serialize().length);
-    const nChange = cCoinControl.nValue - (nFee + nValue);
-    const [changeAddress,changeAddressPath]= await getNewAddress();
-    if (nChange > 0) {
-        // Change output
-        cTx.addoutput(changeAddress, nChange / COIN);
-    } else {
-        // We're sending alot! So we deduct the fee from the send amount. There's not enough change to pay it with!
-        nValue -= nFee;
-    }
-
-    // The primary Standard-to-Cold output
-    const [primaryAddress, primaryAddressPath]= await getNewAddress();
-    cTx.addcoldstakingoutput(primaryAddress, coldAddr, nValue / COIN);
-
-    // Debug-only verbose response
-    if (debug)doms.domHumanReadable.innerHTML = "Balance: " + (nBalance / COIN) + "<br>Fee: " + (nFee / COIN) + "<br>To: " + coldAddr + "<br>Sent: " + (nValue / COIN) + (nChange > 0 ? "<br>Change Address: " + changeAddress + "<br>Change: " + (nChange / COIN) : "");
-
-    // Sign and broadcast!
-    if (hasHardwareWallet()) {
-        // Format the inputs how the Ledger SDK prefers
-        const arrInputs = [];
-        const arrAssociatedKeysets = [];
-        for (const cInput of cTx.inputs) {
-            const cInputFull = await getTxInfo(cInput.outpoint.hash);
-            arrInputs.push([await cHardwareWallet.splitTransaction(cInputFull.hex), cInput.outpoint.index]);
-            arrAssociatedKeysets.push(cInput.path);
-        }
-
-        // Construct the Ledger transaction
-        const cLedgerTx = await cHardwareWallet.splitTransaction(cTx.serialize());
-        const strOutputScriptHex = await cHardwareWallet
-              .serializeTransactionOutputs(cLedgerTx)
-              .toString("hex");
-
-        // Sign the transaction via Ledger
-        createAlert("info", ALERTS.CONFIRM_UNSTAKE_H_WALLET, [{"strHardwareName" : strHardwareName}], 7500);
-        const strSerialisedTx = await cHardwareWallet.createPaymentTransactionNew({
-            inputs: arrInputs,
-            associatedKeysets: arrAssociatedKeysets,
-            outputScriptHex: strOutputScriptHex,
-        });
-        
-	
-        // Broadcast the Hardware (Ledger) tx
-        const result = await sendTransaction(strSerialisedTx, "<b>Delegation successful!</b>");
-        if(result){
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(strSerialisedTx)))))));
-	    
-            if(nChange>0){
-		mempool.addUTXO({id: futureTxid,path: changeAddressPath,sats: nChange,script:Crypto.util.bytesToHex(cTx.outputs[0].script),vout: 0,status: Mempool.PENDING});
-		mempool.addUTXO({id: futureTxid,path: primaryAddressPath,sats: nValue,vout: 1,script:Crypto.util.bytesToHex(cTx.outputs[1].script),status: Mempool.PENDING_COLD});
-            }else{
-		mempool.addUTXO({id: futureTxid,path: primaryAddressPath,script: Crypto.util.bytesToHex(cTx.outputs[0].script),sats: nValue,vout: 0,status: Mempool.PENDING_COLD});
-            }
-        }
-	
-    } else {
-        
-        const sign= await cTx.sign(masterKey, 1);
-	
-        // Broadcast the software TX
-        const result = await sendTransaction(sign, "<b>Delegation successful!</b>");
-        if(result){
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            // Add our delegation + change (if any) to the local mempool
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(sign)))))));
-            if(nChange>0){
-		mempool.addUTXO({id: futureTxid,path: changeAddressPath,sats: nChange,script:Crypto.util.bytesToHex(cTx.outputs[0].script),vout: 0,status: Mempool.PENDING});
-		mempool.addUTXO({id: futureTxid,path: primaryAddressPath,sats: nValue,vout: 1,script:Crypto.util.bytesToHex(cTx.outputs[1].script),status: Mempool.PENDING_COLD});
-            }else{
-		mempool.addUTXO({id: futureTxid,path: primaryAddressPath,script: Crypto.util.bytesToHex(cTx.outputs[0].script),sats: nValue,vout: 0,status: Mempool.PENDING_COLD});
-            }
-        }
-        
-    }
-  doms.domGenIt.innerHTML = "Continue";
-}
-
-// Coin Control response formats
-function ccError    (msg = '') { return { success: false, msg } };
-function ccSuccess  (data)     { return { success: true, ...data } };
-
-function chooseUTXOs(cTx, nTotalSatsRequired = 0, nMinInputSize = 0, fColdOnly = false) {
-    console.log("Constructing TX of value: " + (nTotalSatsRequired / COIN) + " " + cChainParams.current.TICKER);
-
-    // Select the UTXO type bucket
-
-    //const arrUTXOs
-    const arrUTXOs = mempool.UTXOs;
-
-    // Select and return UTXO pointers (filters applied)
-    const cCoinControl = { nValue: 0, nChange: 0, arrSelectedUTXOs: [] }
-    
-    let spent=[]
-    for(let i=0; i< arrUTXOs.length;i++){
-        const cUTXO= arrUTXOs[i]
-        if(!fColdOnly){
-            if(cUTXO.status!==Mempool.CONFIRMED && cUTXO.status!==Mempool.REWARD && cUTXO.status!==Mempool.PENDING){
-		continue;
-            }
-            if(cUTXO.status===Mempool.REWARD && !Mempool.isValidReward(cUTXO)){
-		continue
-            }
-        }
-        if(fColdOnly && cUTXO.status!==Mempool.DELEGATE && cUTXO.status!==Mempool.PENDING_COLD) continue;
-        // Don't spend locked Masternode collaterals
-        if (isMasternodeUTXO(cUTXO)) continue; //CHANGE THIS
-
-        // Have we met the required sats threshold?
-        if (cCoinControl.nValue >= nTotalSatsRequired + getFee(cTx.serialize().length)) {
-            // Required Coin Control value met, yahoo!
-            console.log("Coin Control: TX Constructed! Selected " + cCoinControl.arrSelectedUTXOs.length + " input(s) (" + (cCoinControl.nValue / COIN) + " " + cChainParams.current.TICKER + ")");
-            break;
-        }
-
-        // Does the UTXO meet size requirements?
-        if (cUTXO.sats < nMinInputSize) continue;
-        
-        // Push UTXO and cache new total value
-        cCoinControl.arrSelectedUTXOs.push(cUTXO);
-        cCoinControl.nValue += cUTXO.sats;
-        console.log("Coin Control: Selected input " + cUTXO.id.substr(0, 6) + "(" + cUTXO.vout + ")... (Added " + (cUTXO.sats / COIN) + " " + cChainParams.current.TICKER + " - Total: " + (cCoinControl.nValue / COIN) + ")");
-
-        // Stuff UTXO into the TX
-        cTx.addinput({
-            txid: cUTXO.id,
-            index: cUTXO.vout,
-            script: cUTXO.script,
-            path: cUTXO.path
-        });
-        spent.push(cUTXO);
-    }
-
-    // if we don't have enough value: return false
-    if (cCoinControl.nValue < nTotalSatsRequired) return ccError("Balance is too small! (Missing " + (cCoinControl.nValue - nTotalSatsRequired).toLocaleString('en-gb') + " sats)");
-
-    // Reaching here: we have sufficient UTXOs, calc final misc data and return!
-    cCoinControl.nChange = nTotalSatsRequired - cCoinControl.nValue;
-    return ccSuccess(cCoinControl);
-}
 
 export function isMasternodeUTXO(cUTXO, masternode = null) {
     const cMasternode = masternode || JSON.parse(localStorage.getItem("masternode"));
@@ -1099,150 +792,6 @@ export function isMasternodeUTXO(cUTXO, masternode = null) {
 }
 
 
-async function createTxGUI() {
-    if (!hasWalletUnlocked(true)) return;
-    
-    if(masterKey.isViewOnly) {
-	return createAlert("warning", "Attempting to send funds in view only mode.", 6000);
-    }
-    
-    // Clear the inputs on 'Continue'
-    if (domGenIt.innerHTML === 'Continue') {
-doms.domenIt.innerHTML = 'Send Transaction';
-doms.domxOutput.innerHTML = '';
-doms.domumanReadable.innerHTML = "";
-doms.domalue1s.value = "";
-doms.domddress1s.value = "";
-doms.domeqDesc.value = '';
-doms.domeqDisplay.style.display = 'none';
-	return;
-    }
-    // Sanity check the address
-    const address =doms.domAddress1s.value.trim();
-    // If Staking address: redirect to staking page
-    if (address.startsWith(cChainParams.current.STAKING_PREFIX)) {
-	createAlert('warning', ALERTS.STAKE_NOT_SEND, [], 7500);
-	returndoms.domStakeTab.click();
-    }
-    if (address.length !== 34) return createAlert('warning', ALERTS.BAD_ADDR_LENGTH,[{"addressLength" : address.length}], 2500);
-    if (!cChainParams.current.PUBKEY_PREFIX.includes(address[0])) return createAlert('warning', ALERTS.BAD_ADDR_PREFIX, [{"address" : address[0]},{"addressPrefix" : cChainParams.current.PUBKEY_PREFIX.join(' or ')}], 3500);
-    if (!bitjs.isValidDestination(address,cChainParams.current.PUBKEY_ADDRESS)) return createAlert('warning', ALERTS.INVALID_ADDRESS, [{"address" :address}], 3500);
-    
-    // Sanity check the amount
-    let nValue = Math.round(Number(domValue1s.value.trim()) * COIN);
-    if (nValue <= 0 || isNaN(nValue))  return createAlert('warning', ALERTS.INVALID_AMOUNT +ALERTS.SENT_NOTHING, [], 2500);
-    if (!Number.isSafeInteger(nValue)) return createAlert('warning', ALERTS.INVALID_AMOUNT +ALERTS.MORE_THEN_8_DECIMALS, [], 2500);
-    
-    // Construct a TX and fetch Standard inputs
-    const nBalance = getBalance();
-    const cTx = bitjs.transaction();
-    const cCoinControl = chooseUTXOs(cTx, nValue, 0, false);
-    if (!cCoinControl.success) return alert(cCoinControl.msg);
-    // Compute fee
-    const nFee = getFee(cTx.serialize().length);
-    
-    // Compute change (or lack thereof)
-    const nChange = cCoinControl.nValue - (nFee + nValue);
-    const [changeAddress,changeAddressPath] = await getNewAddress({verify: masterKey.isHardwareWallet});
-
-    const outputs = [];
-    if (nChange > 0) {
-        // Change output
-        outputs.push([changeAddress, nChange / COIN]);
-    } else {
-        // We're sending alot! So we deduct the fee from the send amount. There's not enough change to pay it with!
-        nValue -= nFee;
-    }
-    
-    // Primary output (receiver)
-    outputs.push([address, nValue / COIN]);
-    
-    // Debug-only verbose response
-    if (debug)doms.domHumanReadable.innerHTML = "Balance: " + (nBalance / COIN) + "<br>Fee: " + (nFee / COIN) + "<br>To: " + address + "<br>Sent: " + (nValue / COIN) + (nChange > 0 ? "<br>Change Address: " + changeAddress + "<br>Change: " + (nChange / COIN) : "");
-
-    // Add outputs to the Tx
-    for (const output of outputs) {
-        cTx.addoutput(output[0], output[1]);
-    }
-    
-
-    // Sign and broadcast!
-    if (!masterKey.isHardwareWallet) {
-        const sign = await cTx.sign(masterKey, 1);
-
-        const result = await sendTransaction(sign);
-        // Add our change (if any) to the local mempool
-        if(result){
-            
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(sign)))))));
-            
-            const [isYours,yourPath]= await isYourAddress(address);
-            if(nChange>0){
-		mempool.addUTXO({id: futureTxid,path: changeAddressPath,sats: nChange,script:Crypto.util.bytesToHex(cTx.outputs[0].script) ,vout: 0,status: Mempool.PENDING});
-		if(isYours){
-		    mempool.addUTXO({id: futureTxid,path: yourPath,sats: nValue,vout: 1,script:Crypto.util.bytesToHex(cTx.outputs[1].script),status: Mempool.PENDING});
-		}
-            }else{
-		if(isYours){
-		    mempool.addUTXO({id: futureTxid,path: yourPath,sats: nValue,vout: 0,script:Crypto.util.bytesToHex(cTx.outputs[0].script),status: Mempool.PENDING});
-		}
-            }
-        }  
-    } else {
-        // Format the inputs how the Ledger SDK prefers
-        const arrInputs = [];
-        const arrAssociatedKeysets = [];
-        for (const cInput of cTx.inputs) {
-            const cInputFull = await getTxInfo(cInput.outpoint.hash);
-            arrInputs.push([await cHardwareWallet.splitTransaction(cInputFull.hex), cInput.outpoint.index]);
-            arrAssociatedKeysets.push(cInput.path);
-        }
-        const cLedgerTx = await cHardwareWallet.splitTransaction(cTx.serialize());
-        const strOutputScriptHex = await cHardwareWallet
-              .serializeTransactionOutputs(cLedgerTx)
-              .toString("hex");
-
-        // Sign the transaction via Ledger
-        const strSerialisedTx = await confirmPopup(
-            {
-		title: ALERTS.CONFIRM_POPUP_TRANSACTION,
-		html: createTxConfirmation(outputs),
-		resolvePromise: cHardwareWallet.createPaymentTransactionNew({
-		    inputs: arrInputs,
-		    associatedKeysets: arrAssociatedKeysets,
-		    outputScriptHex: strOutputScriptHex,
-		}),
-            }
-        );
-
-	
-        // Broadcast the Hardware (Ledger) TX
-        const result= await sendTransaction(strSerialisedTx);
-        if(result){
-            for(const tx of cTx.inputs){
-		mempool.autoRemoveUTXO({id: tx.outpoint.hash,path: tx.path,vout: tx.outpoint.index})
-            }
-            // Add our change (if any) to the local mempool
-            const [isYours,yourPath]= await isYourAddress(address);
-            const futureTxid=swapHEXEndian(await hash(Crypto.util.hexToBytes((await hash((Crypto.util.hexToBytes(strSerialisedTx)))))));
-            
-            if(nChange>0){
-		mempool.addUTXO({id: futureTxid,path: changeAddressPath,sats: nChange,script:Crypto.util.bytesToHex(cTx.outputs[0].script) ,vout: 0,status: Mempool.PENDING});        
-		if(isYours){
-		    mempool.addUTXO({id: futureTxid,path: yourPath,sats: nValue,vout: 1,script:Crypto.util.bytesToHex(cTx.outputs[1].script),status: Mempool.PENDING});
-		}
-            }else{
-		if(isYours){
-		    mempool.addUTXO({id: futureTxid,path: yourPath,sats: nValue,vout: 0,script:Crypto.util.bytesToHex(cTx.outputs[0].script),status: Mempool.PENDING});
-		}
-            }
-        }
-    }
-  doms.domGenIt.innerHTML = "Continue";
-}
 
 export async function wipePrivateData() {
     const title = hasEncryptedWallet() ?
@@ -1342,30 +891,6 @@ function addExtraBytes(arrTxBytes, arrPubkeyBytes, nLen) {
     return arrNewTxBytes;
 }
 
-async function createRawTransaction() {
-    // Prepare a TX
-    const cTx = bitjs.transaction();
-    const txid = document.getElementById("prevTrxHash").value;
-    const index = document.getElementById("index").value;
-    const script = document.getElementById("script").value;
-
-    // Primary input
-    cTx.addinput({txid, index, script});
-
-    // Primary output
-    const strAddress = document.getElementById("address1").value;
-    const nValue = document.getElementById("value1").value;
-    cTx.addoutput(strAddress, nValue);
-
-    // Change output
-    const strChange = document.getElementById("address2").value;
-    const nChangeValue = document.getElementById("value2").value;
-    cTx.addoutput(strChange, nChangeValue);
-
-    // Sign via WIF key
-    const wif = document.getElementById("wif").value;
-    document.getElementById("rawTrx").value = await cTx.sign(wif, 1); //SIGHASH_ALL DEFAULT 1
-}
 function addCellToTable(row,data){
     let td=row.insertCell();
     td.appendChild(document.createTextNode(data));
