+++ scripts/index.js
@@ -1,12 +1,16 @@
-import "jquery";
 import "bootstrap";
 import "bootstrap/dist/css/bootstrap.min.css";
 
 import '@fortawesome/fontawesome-free/css/all.css';
 
-import '../assets/style/style.css';
-import { start, openTab, accessOrImportWallet, guiImportWallet, onPrivateKeyChanged, toClipboard, toggleExportUI, wipePrivateData, restoreWallet } from './global.js';
-import { generateWallet, getNewAddress } from "./wallet.js";
-
+import { start } from './global.js';
 window.onload = start;
-export { openTab, generateWallet, accessOrImportWallet, guiImportWallet, onPrivateKeyChanged, getNewAddress, toClipboard, toggleExportUI, wipePrivateData, restoreWallet };
+
+import '../assets/style/style.css';
+
+// Export global functions to the MPW namespace so we can use them in html
+export { openTab, accessOrImportWallet, guiImportWallet, onPrivateKeyChanged, toClipboard, toggleExportUI, wipePrivateData, restoreWallet, refreshChainData } from './global.js';
+export { generateWallet, getNewAddress } from "./wallet.js";
+export { toggleTestnet, toggleDebug }  from "./settings.js";
+export { createTxGUI, createRawTransaction, undelegateGUI, delegateGUI} from "./transactions.js";
+export { hexToBytes, bytesToHex, dSHA256 } from "./utils.js";
